refactor: class component Search rename functional component
