+++ src/components/Serch/search.module.scss
@@ -1,0 +1,87 @@
+@import '../../styles/settings/index';
+
+.search {
+  margin: 24px 0 48px;
+}
+
+.block {
+  @include flexCenterBetween;
+  column-gap: 48px;
+}
+
+.input {
+  border: 3px solid $black;
+  border-radius: 12px;
+  width: 100%;
+  padding: 12px;
+  background-color: $grey;
+  color: $black;
+  font-weight: 700;
+  font-size: 18px;
+  height: 60px;
+  transition: background-color 0.2s ease;
+
+  &:focus {
+    background-color: $white;
+  }
+}
+
+.button {
+  border-radius: 12px;
+  padding: 0 48px;
+  border: 3px solid transparent;
+  background-color: $blue;
+  font-weight: 700;
+  font-size: 20px;
+  height: 60px;
+  @include flexCenter;
+  color: $white;
+  transition:
+    color 0.2s ease,
+    border-color 0.2s ease,
+    background-color 0.2s ease;
+
+  &:not(:disabled) {
+    @include hover {
+      color: $black;
+      border-color: $black;
+      background-color: $white;
+    }
+  }
+
+  &:disabled {
+    opacity: 0.8;
+    cursor: default;
+  }
+}
+
+@media (max-width: $laptop) {
+  .search {
+    margin: 20px 0 40px;
+  }
+
+  .block {
+    column-gap: 40px;
+  }
+
+  .input {
+    height: 50px;
+  }
+
+  .button {
+    height: 50px;
+    padding: 0 40px;
+  }
+}
+
+@media (max-width: $mob) {
+  .block {
+    @include flexColumn;
+    row-gap: 12px;
+  }
+
+  .button {
+    width: 100%;
+    font-size: 20px;
+  }
+}
