+++ src/components/Serch/types.ts
@@ -1,0 +1,7 @@
+interface ISearchProps {
+  searchParams: string;
+  handleSearch: (searchQuery: string) => void;
+  isLoading: boolean;
+}
+
+export type { ISearchProps };
