+++ src/index.css
@@ -1,3 +1,3 @@
 @import './styles/normalize.css';
 @import './styles/fonts.css';
-@import './styles/components/index.scss';
+@import './styles/settings/index.scss';
