--- src/styles/components/_results.scss
@@ -1,15 +1,0 @@
-.results {
-  &__error-btn {
-    background-color: $red;
-    border-radius: 12px;
-    padding: 12px 16px;
-    color: $white;
-    font-size: 16px;
-    transition: opacity 0.2s ease;
-    border: 2px solid $black;
-
-    &:hover {
-      opacity: 0.8;
-    }
-  }
-}
