--- src/styles/components/index.scss
@@ -1,4 +1,0 @@
-@import '../settings/index.scss';
-
-@import './search.scss';
-@import './results.scss';
